Removing the old level up overlay
